Clean up directory structure
Break down db and view into sub-packages for Habit and Habit Event
