+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/db/habitevent/EmbeddedLocation.java
@@ -1,4 +1,4 @@
-package com.github.cmput301f21t44.hellohabits.db;
+package com.github.cmput301f21t44.hellohabits.db.habitevent;
 
 import androidx.room.ColumnInfo;
