+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/view/habit/CreateEditHabitFragment.java
@@ -1,4 +1,4 @@
-package com.github.cmput301f21t44.hellohabits.view;
+package com.github.cmput301f21t44.hellohabits.view.habit;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
