+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/view/habit/HabitAdapter.java
@@ -1,4 +1,4 @@
-package com.github.cmput301f21t44.hellohabits.view;
+package com.github.cmput301f21t44.hellohabits.view.habit;
 
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
@@ -11,19 +11,18 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.github.cmput301f21t44.hellohabits.R;
 import com.github.cmput301f21t44.hellohabits.databinding.LlistHabitItemBinding;
 import com.github.cmput301f21t44.hellohabits.model.Habit;
+import com.github.cmput301f21t44.hellohabits.view.OnItemClickListener;
 
-import java.time.DayOfWeek;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
-import java.util.EnumSet;
-import java.util.Set;
 
 public class HabitAdapter extends ListAdapter<Habit, HabitAdapter.HabitHolder> {
     private final OnItemClickListener<Habit> listener;
-    protected HabitAdapter(@NonNull DiffUtil.ItemCallback<Habit> diffCallback, OnItemClickListener<Habit> listener) {
+
+    public HabitAdapter(@NonNull DiffUtil.ItemCallback<Habit> diffCallback, OnItemClickListener<Habit> listener) {
         super(diffCallback);
         this.listener = listener;
     }
@@ -58,8 +57,6 @@ public class HabitAdapter extends ListAdapter<Habit, HabitAdapter.HabitHolder> {
 
     protected static class HabitHolder extends RecyclerView.ViewHolder {
         private final LlistHabitItemBinding mItemBinding;
-        private static final DateTimeFormatter formatter =
-                DateTimeFormatter.ofPattern("d MMMM, y").withZone(ZoneId.systemDefault());
 
         HabitHolder(@NonNull LlistHabitItemBinding itemBinding) {
             super(itemBinding.getRoot());
@@ -70,51 +67,17 @@ public class HabitAdapter extends ListAdapter<Habit, HabitAdapter.HabitHolder> {
             mItemBinding.titleView.setText(habit.getTitle());
             mItemBinding.reasonView.setText(habit.getReason());
             // check level of consistency
-            Instant instant = habit.getDateStarted();
-            LocalDate startDate = instant.atZone(ZoneId.of( "America/Edmonton" )).toLocalDate();
-            LocalDate endDate = LocalDate.now();
-            int startDay = startDate.getDayOfWeek().getValue();
-            int endDay = endDate.getDayOfWeek().getValue();
-            long days = ChronoUnit.DAYS.between(startDate, endDate);
-            boolean recurranceDays[] = habit.getDaysOfWeek();
-            int numRecurranceDays = 0;
-            for (int i = 0;i<recurranceDays.length;i++){
-                if(recurranceDays[i]){
-                    numRecurranceDays++;
-                }
-            }
-            long totalDays = 0;
-            for (int i = startDay;i<=7;i++){
-                if (recurranceDays[i-1]){
-                    totalDays++;
-                }
-                days--;
-            }
-            for(int i = endDay;i>=1;i--){
-                if (recurranceDays[i-1]){
-                    totalDays++;
-                }
-                days--;
-            }
-            totalDays = totalDays + numRecurranceDays*days/7;
-            double consistency = 1;
-            if(totalDays!=0) {
-                consistency = habit.getEvents().size() / totalDays;
-            }
-            // if consistency <50% red logo, if between 50 and 75%, yellow logo,
-            // if greater than 75% green logo.
-            // if habit is non repetitive or hasnt started yet, consistency = 100%
-            if(consistency<0.5){
+            double consistency = Habit.getConsistency(habit);
+
+            if (consistency < 0.5) {
                 mItemBinding.imageView.setImageResource(R.mipmap.red_logo_round);
-            }
-            else if(consistency<0.75){
+            } else if (consistency < 0.75) {
                 mItemBinding.imageView.setImageResource(R.mipmap.yellow_logo_round);
-            }
-            else{
+            } else {
                 mItemBinding.imageView.setImageResource(R.mipmap.green_logo_round);
             }
 
-            mItemBinding.getRoot().setOnClickListener(v-> listener.onItemClick(habit));
+            mItemBinding.getRoot().setOnClickListener(v -> listener.onItemClick(habit));
         }
 
     }
