+++ app/src/main/res/layout/fragment_create_edit_habit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CreateEditHabitFragment">
+    tools:context=".view.habit.CreateEditHabitFragment">
 
     <Button
         android:id="@+id/button_back"
