+++ app/src/main/res/layout/fragment_create_edit_habit_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CreateEditHabitFragment"
+    tools:context=".view.habit.CreateEditHabitFragment"
     tools:layout_editor_absoluteX="-4dp"
     tools:layout_editor_absoluteY="-16dp">
