+++ app/src/main/res/layout/fragment_todays_habits.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.TodaysHabitsFragment">
+    tools:context=".view.habit.TodaysHabitsFragment">
 
     <Button
         android:id="@+id/button_new_habit"
